Cleanup unneeded functions
Rename files to match their names in libyuv with `libyuv_mini` prefix added

Put comment in CMakeLists.txt that states the commit from where libyuv files were imported
